pyproject.toml
[project]
name = "otee-diagnostic"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.128.0",
    "google-genai>=1.60.0",
    "google-generativeai>=0.8.6",
    "loguru>=0.7.3",
    "lxml>=6.0.2",
    "openai>=2.15.0",
    "pydantic-settings>=2.12.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
